pyproject.toml
[tool.poetry]
name = "ztcodewars"
version = "0.1.0"
description = ""
authors = ["Phillip Samuel K. <phillipsamuelk3@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.1"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
emoji = "^2.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
